Split `appChanges.ts` out of `apps.ts`
